refactor(domain): 패키지 구조 변경
- DriverDecision을 util에서 domain으로 패키지 변경
- 숫자 생성관련 파일들을 numbergenerator 패키지로 변경
