+++ src/main/kotlin/domain/DriverDecision.kt
@@ -1,6 +1,6 @@
-package util
+package domain
 
 enum class DriverDecision(val distance: String) {
     GO("-"),
-    STOP("");
+    STOP(""),
 }
