+++ src/main/kotlin/domain/numbergenerator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.numbergenerator
 
 interface NumberGenerator {
     fun generate(): Int
