+++ src/main/kotlin/domain/numbergenerator/RandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.numbergenerator
 
 import kotlin.random.Random
