+++ src/test/kotlin/domain/DriverTest.kt
@@ -3,7 +3,6 @@ package domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
-import util.DriverDecision
 import util.TestNumberGenerator
 
 internal class DriverTest {
