+++ src/test/kotlin/util/TestNumberGenerator.kt
@@ -1,6 +1,6 @@
 package util
 
-import domain.NumberGenerator
+import domain.numbergenerator.NumberGenerator
 
 class TestNumberGenerator(private val testNumber: Int) : NumberGenerator {
     override fun generate() = testNumber
